refactor: move format duration ms to utils
